backend/pyproject.toml
[tool]
[project]
name = "project-einstein-backend"
version = "0.1.0"
description = "FastAPI service for Project Einstein"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"fastapi>=0.111,<1.0",
	"uvicorn[standard]>=0.30,<1.0",
	"pydantic>=2.7,<3",
	"pydantic-settings>=2.4,<3",
	"python-multipart>=0.0.9",
	"pillow>=10,<12",
	"httpx>=0.28.1",
	"openai>=2.3.0",
	"pytest>=8.4.2",
	"pyyaml>=6.0.3",
]

[tool.pytest.ini_options]
pythonpath = ["."]
